moved swap_uint64 to separate file
